Add V2 logo variations, lossless and lossy optimized
This adds several new variants of the logo:

- An inverted artichoke logo (green on black).
- Navbar branding logos.
- Artichoke mark-only logos.
- Wordmark logos with mark and "Artichoke Ruby".
- A social profile photo (SVG and PNG) with white background that is able
  to be cropped in a circle.
- The social banner image has a "not quite black" background.

SVGs in `v2/img` are cleaned with SVGO. SVGs in `v2/optimized` are
cleaned and minified with SVGO. PNGs in `v2/img` are losslessly minified
with pngcrush. PNGs in `v2/optimized` are lossily compressed from the
pngcrushed images in `v2/img` with pngquant.